# CAR-2259: Faster VM import and export