--- a/ocaml/test/OMakefile
+++ b/ocaml/test/OMakefile
@@ -1,6 +1,6 @@
 OCAMLPACKS = oUnit sexpr xcp xmlm stunnel xml-light2 http-svr uuid	\
              netdev tapctl xenctrl xenctrlext xenstore-compat	\
-             pciutil oclock gzip sha1 xcp.network xcp.rrd xcp.storage	\
+             pciutil oclock gzip sha1 sha.sha1 xcp.network xcp.rrd xcp.storage	\
              xcp.xen xcp.memory tar tar.unix oPasswd xcp-inventory \
              rrdd-plugin pci xapi-test-utils