--- a/ocaml/xapi/stream_vdi.ml
+++ b/ocaml/xapi/stream_vdi.ml
@@ -262,10 +262,11 @@ let recv_all refresh_session ifd (__context:Context.t) rpc session_id vsn force 
 		  done
 	      end;
 		
-	    let csum = Sha1sum.sha1sum
-	      (fun checksumfd ->
-		 Tar_unix.Archive.multicast_n ifd [ ofd; checksumfd ] length) in
-	    
+	    let buffer = String.make (Int64.to_int length) '\000' in
+	    Unixext.really_read ifd buffer 0 (Int64.to_int length);
+	    Tar_unix.Archive.multicast_n_string buffer [ ofd ] (Int64.to_int length);
+	    let csum = Sha1.to_hex (Sha1.string buffer) in
+	
 	    checksum_table := (file_name, csum) :: !checksum_table;
 
 	    Tar_unix.Archive.skip ifd (Tar_unix.Header.compute_zero_padding_length hdr);