This patch series adds some vhd-tool VDI import and export performance features to Xapi to improve its VM import and export performance. This is considered a better option than calling vhd-tool from Xapi, because adding Xapi VM import and export features to vhd-tool is considerably more difficult and error-prone.

The resulting VM import and export performance in Xapi is better than the corresponding VDI import and export performance in vhd-tool, for the same amount of VDI data:

| Time for a 10GiB VDI with random data | `xe vm-export`                         | `vhd-tool export` | `xe vm-import`                         | `vhd-tool import` | 
|---------------------------------------|----------------------------------------|-------------------|----------------------------------------|-------------------|
| before patches                        | 131s (sha1sum)                         | 68s (ocaml-sha)   | 102s (sha1sum)                         | 66s (ocaml-sha)   |  
| after patches                         |  60s **(faster than vhd-tool export)** |                   |  56s **(faster than vhd-tool import)** |                   |