@euanh 2016-08-18
@mg12 You need to add a dependency on the `sha` package in the `opam` file.