@euanh 2016-08-18
@mg12 You need to depend on `sha` in the `opam` file, not `sha1`.